requirements.txt
dm_control==1.0.12
ale_py==0.8.1
gym==0.26.2
kornia==0.6.12
matplotlib==3.6.2
numpy==1.24.3
opencv_python==4.7.0.72
pandas==2.0.1
pytorch_lightning==2.0.2
PyYAML==6.0
torchvision=0.15.2
torch==2.0.1
tensorboard==2.13.0